+++ docs/fr_FR/changelog.md
@@ -4,6 +4,12 @@ Liste des versions du plugin googlecast.
 
 *[Retour à la documentation](index.md)*
 
+## Version du 08 janvier 2020 (beta) - 2.13
+
+- suppression du moteur 'TTS Google Speech API', remplacé par 'Google Cloud Text-To-Speech' (https://cloud.google.com/text-to-speech/)
+(voir FAQ documentation si la clé API ne fonctionne plus)
+- mise à jour de la documentation
+
 ## Version du 07 janvier 2020 (beta) - 2.12
 
 - optimisations et corrections de bugs